+++ wdio/features/step-definitions/login-view.js
@@ -7,17 +7,5 @@ Then(/^Login screen is displayed/, async () => {
 });
 
 When(/^I tap Reset Wallet on Login screen/, async () => {
-    await LoginScreen.tapResetWallet();
-});
-
-When(/^I tap I understand, continue on Delete wallet modal/, async () => {
-    await LoginScreen.tapIUnderstandContinue();
-});
-
-When(/^I type "([^"]*)?" on Delete wallet modal permanently/, async (text) => {
-    await LoginScreen.typedelete(text);
-});
-
-When(/^I tap Delete my wallet on Delete wallet modal permanently/, async () => {
-    await LoginScreen.tapDeleteMyWallet();
+    await LoginScreen.tapResetWalletButton();
 });
