+++ wdio/features/testIDs/Components/DeleteWalletModal.testIds.js
@@ -1,0 +1,4 @@
+export const DELETE_MODAL_UNDERSTAND_CONTINUE_ID = 'delete-modal-understand-continue';
+export const DELETE_MODAL_CANCEL_BUTTON = 'delete-modal-cancel-button';
+export const DELETE_MODAL_DELETE_INPUT_ID = 'delete-wallet-inputbox';
+export const DELETE_MODAL_DELETE_MY_WALLET_PERMANENTLY = 'delete-my-wallet-button-on-permanetly-delete-modal';
