+++ wdio/features/testIDs/Screens/LoginScreen.testIds.js
@@ -1,5 +1,1 @@
-export const LOGIN_VIEW_RESET_WALLET_ID = 'login-view-reset-wallet';
-export const DELETE_MODAL_UNDERSTAND_CONTINUE_ID = 'delete-modal-understand-continue';
-export const DELETE_MODAL_CANCEL_BUTTON = 'delete-modal-cancel-button';
-export const DELETE_MODAL_DELETE_INPUT_ID = 'delete-wallet-inputbox';
-export const DELETE_MODAL_DELETE_MY_WALLET_PERMANENTLY = 'delete-my-wallet-button-on-permanetly-delete-modal';
+export const LOGIN_VIEW_RESET_WALLET_ID = 'reset-wallet-button';
